Make schema-view-section dir=ltr
Move the dir=ltr from the schema-text to the schema-view-section, so
that we don’t have to change the grid layout to keep the links to the
right side of the schema text (because the schema text is always LTR).
To put the links themselves in the right direction again, add another
dir attribute there based on the language code for messages.

Bug: T369283
Change-Id: If7759aac6e274c3b68eee945fd2c3deae0acd2a5