Temporarily disable IP Masking (temporary accounts) in CI
This lets us implement proper support for IP Masking in n>1 commits
without CI being broken during the intermediate commits. It’s a bit
ugly, and needs a few fixes in other extensions that already started to
assume that temporary accounts are enabled in CI, but it’s not too bad.

Bug: T356148
Bug: T370577
Change-Id: Ibd020a526055cf355902869ecf6b993ee4c67c95
Depends-On: I69746301c5cd9c8908271a40f8083e039beca6d2
Depends-On: Ib552a7d41ce3625a5ae00ad9137b2fe2becae0e2