Use EntitySchemaStatus in EntitySchemaUpdater
Use a Status object to collect and propagate all errors, mostly using
the same messages as before. This will make it easier to add IP Masking
support later.

EditConflict is gone; instead, EntitySchemaUpdater always returns an
'edit-conflict' status in that case, which EntitySchemaEditAction and
SetEntitySchemaLabelDescriptionAliases then replace with a more specific
error message if applicable. (The EditConflict previously thrown by
overwriteWholeSchema() wasn’t caught at all in UndoSubmitAction /
RestoreSubmitAction, as far as I can tell; now it would be reported with
the generic 'edit-conflict' message – though I’m not sure if it actually
happens: I haven’t been able to reproduce it locally.)

In EntitySchemaUpdateGuard, propagate the underlying PatcherException to
signal an edit conflict; this isn’t super pretty, but to me it seems
slightly preferable to returning a blank Status (with type-unsafe
PersistenceEntitySchemaData inside), creating a new Status subclass just
for this one method, or keeping EditConflict around just for this
method. But I wouldn’t mind changing this either.

Remove the mentions of InvalidArgumentException – unchecked exceptions
don’t need to be documented [1]. (Note that DomainException, which
according to testUpdateSchemaText_throwsForUnknownSerializationVersion()
may also be thrown, wasn’t documented either.)

[1]: https://www.mediawiki.org/wiki/Manual:Coding_conventions/PHP#Exception_handling

Bug: T356148
Change-Id: I9743b2cacc0c0b333ab355229d4e68dc205f0f6b