Make WatchlistUpdater a service
Inject the two services it needs properly, while passing the affected
User into the specific methods. The latter part is required for IP
Masking, because the correct User in that case is a temporary account
that was just created, not the original context user.

The namespace number can just be hard-coded – nobody was passing in a
different value anyway.

Bug: T356148
Change-Id: Icfab8e5e5aeb1b057cf874f078d2f1b9febc59b2