Support temporary accounts
EntitySchemaStatus and PageUpdaterStatus now extend TempUserStatus (from
Wikibase), containing not just the EntitySchema ID but also the created
temp user and new context (if applicable). MediaWikiPageUpdaterFactory
is responsible for allocating the temp user, if necessary, and returns
the result in the PageUpdaterStatus; the schema inserter and updater
then propagate this through to the EntitySchemaStatus they return. The
various endpoints that make edits (special pages and actions) then
redirect to the resulting EntitySchema, possibly first going through
loginwiki according to the TempUserCreatedRedirect hook, via the shared
EntitySchemaRedirectTrait.

The code to set the context title to the EntitySchema title for the edit
filter hook, previously duplicated between the inserter and the updater,
can now also live in the MediaWikiPageUpdaterFactory, since we have to
create a new context there anyway (for the new temp user).

Bug: T356148
Change-Id: I6f193987f238438265dc46a2e0482516affe5316