Clean up MediaWikiRevisionEntitySchemaUpdaterTest
Extract $services instead of getting them repeatedly (each call to
getServiceContainer() not only gets the test-local services but also
compares them to the global ones, so it’s not a super cheap call), and
fix the typo “schmea” (for “schema”) everywhere while we’re at it.

This change became unnecessary as I worked on T356148 (initially, I had
added a service to the MediaWikiRevisionEntitySchemaUpdater, but in the
end I injected the TempUserCreator into the MediaWikiPageUpdaterFactory
instead), so I moved it to the end of the chain. But I think it’s still
a nice improvement.

Bug: T356148
Change-Id: I65ec70524c27eb57c241a1e9f6198a91cba8184b