Use TempUserTestTrait instead of mocking TempUserConfig
Once these tests (or rather, the tested code) actually start to create
temp users (change I6f193987f2), we’ll need this trait anyway, and it
makes sense to then use the real TempUserConfig instead of a mock as
well; but we can also do this separately. IMHO it makes the code
slightly more readable too.

Bug: T356148
Change-Id: Id46695a94989411a7ae43e79d3676a28c46496ee