Make LoadExtensionSchemaUpdatesHookHandlerTest work on client-only
The hook handler skips itself if EntitySchema repo features are disabled
or if WikibaseRepository is not installed, so let’s make a separate test
for each condition; one test will be skipped in the main CI build, the
other in the client-only build.

Implemented using markTestSkippedIfExtensionNotLoaded() which we inherit
from MediaWikiIntegrationTestCase; for the other test, we define
markTestSkippedIfExtensionLoaded() ourselves. (We could upstream the
latter, but I’m not sure it’s worth it; codesearch [1] only finds one
other extension, SemanticMediaWiki, which appears to skip tests if a
certain other extension is installed.)

[1]: https://codesearch.wmcloud.org/search/?q=if+%5C%28+%28ExtensionRegistry%3A%3AgetInstance%5C%28%5C%29%7C.*Registry%29-%3EisLoaded%5C%28&files=Test&excludeFiles=&repos=

Change-Id: I26997e8efa9768a27410f54472cd836ca3a1c877