Define search fields for CirrusSearch
They’re not actually populated with data yet, that will come later.

Note that, contrary to the original plan in T376250, we have to reuse
the fields from WikibaseCirrusSearch – trying to introduce our own
copies of them would blow up the index with too many additional fields.
I *think* the code should work even without WikibaseCirrusSearch (PHP
apparently allows undefined classes in type hints), so WBCS is only an
optional dependency (though we log a warning if it’s missing and the
search engine is CirrusSearch, because that’s probably not a very useful
setup).

Bug: T376250
Change-Id: I7d2d06f8ea43e6fe57d9a0559b1f00579abd45b1