Add labels data for CirrusSearch
We delegate this almost entirely to WikibaseCirrusSearch – just return
whatever data its fields return. But in order to be able to call it, we
need to make EntitySchema data available as a LabelsProvider (and
AliasesProvider); a new class, SearchEntitySchemaAdapter, is used for
that (created by EntitySchemaConverter). Note that this should only be
used for the search.

In testGetDataForSearchIndex(), we want to avoid depending on
WikibaseCirrusSearch internals, so we inject our own mock field
definitions. (The test still requires the WikibaseCirrusSearch extension
to be loaded so that the LabelsProviderFieldDefinitions class exists.)

Bug: T376251
Change-Id: I8e33e05456c5252167c12e29c2b59129d41b70e8
Depends-On: Ic40ed179bb9e93b9e868874dad997af01a7af5ff