Add fallback language support to EntitySchema page headings
EntitySchema currently displays `No label defined` as the heading
on entity schema pages if no label exists for the EntitySchema in
the current user language. Elsewhere in Wikibase we use fallbacks
to display labels in alternative languages of the users preference
where those labels are available.

Implement the language fallback support for the EntitySchema page
headings.

Bug: T228423
Change-Id: I16f962fbb658702cf2476b4813836e96e55c4931