Strip fallback indicator from HTML page titles
The `EntitySchemaSlotViewRenderer` uses `setDisplayTitle` to set
the title of the EntitySchema pages. This in turn sets the HTML
title of the page to the HTML-stripped version of the same title.
Where the title includes a language fallback indicator, we do not
want this to be included in the plain text of the HTML title.

Override the View action and execute a custom
`ViewEntitySchemaAction` with a `show` implementation that
overrides the HTML Title with a text-only version of the title
(where the fallback indicator is absent) which has been stashed
in the extension metadata.

The Last-Modified-cache handling has been copied here from
`RawAction`. It would have been nice to have a way to reuse the
code from `ViewAction` without needing to extend the class (it is
not stable to extend) or wrap it and delegate (which would
require us to have access to another equivalent `ViewAction`
object to which to delegate), but the current structure of the
classes does not make this possible.

T287713 and T287650 suggest that making `ViewAction` stable to
extend is not a desirable solution here, so this approach has
not been pursued further.

Bug: T228423
Change-Id: I18172e52bd771152a49959b4de43d85c16b6e6d6