Strip fallback indicator from HTML page titles
The `EntitySchemaSlotViewRenderer` uses `setDisplayTitle` to set
the title of the EntitySchema pages. This in turn sets the HTML
title of the page to the HTML-stripped version of the same title.
Where the title includes a language fallback indicator, we do not
want this to be included in the plain text of the HTML title.

Intercept the View action and strip any <sup> tags from the title
before the <title> tag is sent.

Bug: T228423
Change-Id: I18172e52bd771152a49959b4de43d85c16b6e6d6