Avoid PHP Notice on missing entityschema-meta-tags
The ParserOutput / OutputPage can be missing the entityschema-meta-tags
in two cases: when the parser output is stale (from before change
I18172e52bd / commit 567ac71f6c was deployed), or when viewing a page
that doesn’t exist. In both cases, it’s enough to check that the value
is null and simply not do anything in that case; the default title is
good enough in those cases, as far as I can tell from local testing.

The test has quite a bit of copy+paste from the other test in that file,
but I’m not convinced it’s worth extracting a shared setup method yet.

Bug: T385272
Change-Id: I7615e56820fc198a9514173a325840afcd790eb8
(cherry picked from commit f62c7734169b9e6f23264a64eb47426f8e3b2c93)