Fix `PermissionManager::getCascadeProtectionSources` mock
Change Ia5863f418538106f4fd657c672298ff6ac835805 modifies the
public interface of `PermissionManager` such that
`getCascadeProtectionSources` now returns (and is required to
return) a four-element array instead of a two-element array. This
was causing secondary CI for EntitySchema to change.

Update the `PermissionManager` mock in `EntitySchemaEditActionTest`
to match the new `getCascadeProtectionSources` return type.

Bug: T386608
Change-Id: Ic3fb28a78acc87672adcdf5e55cb4d21095ce5ab