Register `WikibaseRepoSearchableEntityScopes` hook and register scope
I10d4c8d9838a94bbdc3f35cc8961f15f957d6b54 adds a new hook to Wikibase
that allows extensions to register additional scopes for the scoped
typeahead search select dropdown. We want Entity Schemas to appear in
the dropdown, so we need to implement the hook.

Register for the hook and add the EntitySchema namespace to the list
of available scopes for the scoped typeahead search dropdown.

Depends-On: I10d4c8d9838a94bbdc3f35cc8961f15f957d6b54
Bug: T387986
Change-Id: I9d0345d40d6085fa2d3d59f76476a78c50cd15ac