Rename propertyType from EntitySchema to WikibaseEntitySchema
Previously, the wikibase:propertyType value for properties with data
type entity-schema defaulted to wikibase:EntitySchema (generated by
RdfVocabulary::getDataTypeURI() by transforming the kebab-case internal
data type name into UpperCamelCase). As outlined in T371196#10347536,
for consistency we want it to be wikibase:WikibaseEntitySchema instead,
so that wikibase:EntitySchema can one day be the class of EntitySchemas
themselves instead.

To be able to control this rollout in production, put it behind a
temporary feature flag.

Bug: T371196
Change-Id: I1d9da2f6fe2792e4e634f901f07d28e7c00527c9